requirements.txt
attrs==25.3.0
branca==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
folium==0.20.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2025.4.0
